requirements.txt
fastapi
uvicorn[standard]
python-multipart
sqlalchemy[asyncio]
asyncpg
psycopg2-binary
langchain
langchain-openai
langchain-core
langchain-community
langgraph
openai
pydantic
pydantic-settings
jsonschema
typing-extensions
structlog
python-json-logger
httpx
requests
python-dateutil
pytest
pytest-asyncio
nest-asyncio
aiofiles
tenacity
python-dotenv
